WIP reproducible notebooks
